% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_system_attributes}
\alias{add_system_attributes}
\title{Add system attributes to an R object}
\usage{
add_system_attributes(x)
}
\arguments{
\item{x}{an R object}
}
\value{
an object of the same time with system attributes attached
}
\description{
For long jobs, it is helpful to retain various system information
in simulation output.  We retain the output of
\code{\link{Sys.info}}, \code{\link{Sys.getpid}},
\code{\link{Sys.time}}, \code{\link{getwd}},
\code{\link{sessionInfo}}, and the value of
\code{\link{R.version}}.
}
