% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{all_the_same}
\alias{all_the_same}
\title{Are all the components of an object identical?}
\usage{
all_the_same(x)
}
\arguments{
\item{x}{an R object}
}
\value{
logical
}
\description{
Are all the components of an object identical?
}
\details{
This is used in next generation functions to determine if
    an evolving population has fixed.
}
\seealso{
\code{\link{selection_step_WF}}, \code{\link{selection_step_M}},
    \code{\link{selection_step_BD}}
}
