% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{benefit}
\alias{benefit}
\title{Benefit function for \code{snowdrift.game}}
\usage{
benefit(model, tau)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{tau}{total group contribution}
}
\description{
Benefit function for \code{snowdrift.game}
}
\seealso{
\code{\link{dbenefit}}
}
