% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{c2_fun}
\alias{c2_fun}
\title{Quadratic cost coefficient}
\usage{
c2_fun(Npop, nPlayers, nu, xi)
}
\arguments{
\item{Npop}{population size}

\item{nPlayers}{number of players}

\item{nu}{non-negative real number that indexes the examples}
}
\description{
Quadratic cost coefficient
}
\seealso{
\code{\link{setup_quadratic_snowdrift_game}}
}
