% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{calculate_derived_parms}
\alias{calculate_derived_parms}
\title{Calculate derived parameters for a \code{snowdrift.game} model}
\usage{
calculate_derived_parms(model)
}
\arguments{
\item{model}{a \code{snowdrift.game} object}
}
\description{
Calculate derived parameters for a \code{snowdrift.game} model
}
\seealso{
\code{\link{create_snowdrift_game}}
}
