% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{compute_fitnesses}
\alias{compute_fitnesses}
\title{Calculate fitness for each member of a population playing a
\code{\link{snowdrift.game}}}
\usage{
compute_fitnesses(model, pop)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{pop}{population matrix; FIX: document this in one place}
}
\value{
population matrix
}
\description{
Calculate fitness for each member of a population playing a
\code{\link{snowdrift.game}}
}
