% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{compute_fixation_stats}
\alias{compute_fixation_stats}
\title{Compute fixation statistics}
\usage{
compute_fixation_stats(
  x,
  nReplicates,
  mutFixCount,
  didNotFixCount,
  mutFixTime,
  resFixTime
)
}
\arguments{
\item{x}{a single fixation record}
}
\description{
Compute fixation statistics
}
