% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{compute_transition_prob_mat}
\alias{compute_transition_prob_mat}
\title{Calculate transition probabilities for birth-death process defined
in invasion/fixation ms.}
\usage{
compute_transition_prob_mat(fitDiff)
}
\arguments{
\item{fitDiff}{vector of fitness differences}
}
\description{
If the fitness difference is calculated between type A and type B
(\code{W_A-W_B}), then the state of the population is the number of
individuals of type A in the population.
Output is a matrix of 3 columns and \code{(#states-1)} rows, labelled
\code{"down"}, \code{"up"} and \code{"stay"}.
    The i'th entry of the column labelled "up" is the transition probability from state i to state i+1.
    The i'th entry of the column labelled "down" is the transition probability from state i to state i-1.
    The i'th entry of the column labelled "stay" is the probability to remain at state i,
       which is 0 for this process.
The \code{"stay"} column is included for compatibility with the function NewGenerationBD.
This definition of the state of the population means that if \code{fitDiff>0}, then
\code{output["up"]>output["down"]}.
}
\details{
This function can process a vector \code{fitDiff}.  For
    one population state, \code{fitDiff} is the mean fitness
    difference between mutants and residents, and thus a
    scalar. However, inputing the vector of fitness differences for
    all mixed states is useful (see
    \code{\link{MutantFixProbExactBD}}).
}
\note{
FIX: this function ought to take another argument, the
    transition probability ratio parameter \eqn{\phi}, with default
    \code{1}.  See Appendix A of invasion-fixation ms.  More
    generally, the implicit assumption that \eqn{F(x)=e^{-\phi x}}
    is unnecessary.  This function could take a function as
    argument, with \eqn{e^{-x}} as the default:
    \code{function(fitDiff, tpr_fun=function(x){exp(-x)})}
}
