% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sigmoid.R
\name{convert_Cornforth_to_David}
\alias{convert_Cornforth_to_David}
\title{Convert original (Cornforth et al. 2012) sigmoid parametrization to
David's parametrization.}
\usage{
convert_Cornforth_to_David(a, b, beta, k, d)
}
\arguments{
\item{a, b, beta, k, d}{Cornforth et al parameters}
}
\description{
Convert original (Cornforth et al. 2012) sigmoid parametrization to
David's parametrization.
}
\note{
I didn't write this conversion up anywhere.
}
\examples{
pDavidV1 <- convert_Cornforth_to_David(a=100, b=0.2, beta=4.76, k=10, d=1)
pTauMMF <- convert_Cornforth_to_tau_mmf(a=100, b=0.2, beta=4.76, k=10, d=1)
pDavidV2 <- convert_tau_mmf_to_David(tauturn=pTauMMF[["tauturn"]], dtau=pTauMMF[["dtau"]], mmf=pTauMMF[["mmf"]])
}
\references{
\insertRef{Corn+12}{evolvr}
}
