% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sigmoid.R
\name{convert_Cornforth_to_L_tauturn_mmf}
\alias{convert_Cornforth_to_L_tauturn_mmf}
\title{Convert original (Cornforth et al. 2012) sigmoid parametrization to
parametrization using the asymptotic benefit \code{L},
the maximal marginal fitness \code{mmf}, and the inflection point \code{tauturn}.}
\usage{
convert_Cornforth_to_L_tauturn_mmf(a, b, beta, k, d)
}
\arguments{
\item{a, b, beta, k, d}{Cornforth et al parameters}
}
\description{
Convert original (Cornforth et al. 2012) sigmoid parametrization to
parametrization using the asymptotic benefit \code{L},
the maximal marginal fitness \code{mmf}, and the inflection point \code{tauturn}.
}
\note{
I didn't write this conversion up anywhere.
}
\examples{
pDavidV1 <- convert_Cornforth_to_David(a=100, b=0.2, beta=4.76, k=10, d=1)
pLTauturnMMFV1 <- convert_Cornforth_to_L_tauturn_mmf(a=100, b=0.2, beta=4.76, k=10, d=1)
pDavidV2 <- convert_L_tauturn_mmf_to_David(L=pLTauturnMMFV1[["L"]], tauturn=pLTauturnMMFV1[["tauturn"]], mmf=pLTauturnMMFV1[["mmf"]])
pLTauturnMMFV2 <- convert_David_to_L_tauturn_mmf(A=pDavidV1[["A"]], K=pDavidV1[["K"]], L = pDavidV1[["L"]])
}
\references{
\insertRef{Corn+12}{evolvr}
}
