% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sigmoid.R
\name{convert_David_to_DC}
\alias{convert_David_to_DC}
\title{Converts David's sigmoid parametrization to DC-style.}
\usage{
convert_David_to_DC(A, K, L)
}
\arguments{
\item{A, K, L}{David's parameters}
}
\description{
Converts David's sigmoid parametrization to the maximizing total good
\code{taumax}, the payoff extrema difference, \code{PED}, and the maximal
marginal fitness \code{mmf}.
}
\note{
See easierMS appendix A.3 for derivation.
}
\examples{
pDavidV1 <- convert_Cornforth_to_David(a=100, b=0.2, beta=4.76, k=10, d=1)
pDC <- convert_David_to_DC(A=pDavidV1[["A"]], K=pDavidV1[["K"]], L = pDavidV1[["L"]])
pDavidV2 <- convert_DC_to_David(taumax=pDC[["taumax"]], ped=pDC[["ped"]], mmf=pDC[["mmf"]])
}
\references{
\insertRef{Corn+12}{evolvr}
}
