% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sigmoid.R
\name{convert_L_tauturn_mmf_to_David}
\alias{convert_L_tauturn_mmf_to_David}
\title{Convert parametrization based on the
asymptotic benefit \code{L}
inflection point \code{tauturn},
and the maximal marginal fitness \code{mmf} to David's sigmoid parametrization.}
\usage{
convert_L_tauturn_mmf_to_David(L, tauturn, mmf)
}
\arguments{
\item{L}{asymptotic benefit}

\item{tauturn}{inflection point}

\item{mmf}{maximal marginal fitness}
}
\description{
Convert parametrization based on the
asymptotic benefit \code{L}
inflection point \code{tauturn},
and the maximal marginal fitness \code{mmf} to David's sigmoid parametrization.
}
\examples{
pDavidV1 <- convert_Cornforth_to_David(a=100, b=0.2, beta=4.76, k=10, d=1)
pLTauturnMMFV1 <- convert_Cornforth_to_L_tauturn_mmf(a=100, b=0.2, beta=4.76, k=10, d=1)
pDavidV2 <- convert_L_tauturn_mmf_to_David(L=pLTauturnMMFV1[["L"]], tauturn=pLTauturnMMFV1[["tauturn"]], mmf=pLTauturnMMFV1[["mmf"]])
pLTauturnMMFV2 <- convert_David_to_L_tauturn_mmf(A=pDavidV1[["A"]], K=pDavidV1[["K"]], L = pDavidV1[["L"]])
}
\references{
\insertRef{Corn+12}{evolvr}
}
