% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sigmoid.R
\name{convert_tau_mmf_to_David}
\alias{convert_tau_mmf_to_David}
\title{Convert sigmoid parametrization based on the inflection point
tauturn, the distance between taumin and taumax and the maximal
marginal fitness mmf, to David's parametrization.}
\usage{
convert_tau_mmf_to_David(tauturn, dtau, mmf)
}
\arguments{
\item{tauturn}{inflection point}

\item{dtau}{distance between \code{taumin} and \code{taumax}}

\item{mmf}{maximal marginal fitness}
}
\description{
Convert sigmoid parametrization based on the inflection point
tauturn, the distance between taumin and taumax and the maximal
marginal fitness mmf, to David's parametrization.
}
\note{
See easierMS appendix A.3 for derivation.
}
\examples{
pDavidV1 <- convert_Cornforth_to_David(a=100, b=0.2, beta=4.76, k=10, d=1)
pTauMMF <- convert_Cornforth_to_tau_mmf(a=100, b=0.2, beta=4.76, k=10, d=1)
pDavidV2 <- convert_tau_mmf_to_David(tauturn=pTauMMF[["tauturn"]], dtau=pTauMMF[["dtau"]], mmf=pTauMMF[["mmf"]])
}
\references{
\insertRef{Corn+12}{evolvr}
}
