% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sigmoid.R
\name{convert_taumax_to_David}
\alias{convert_taumax_to_David}
\title{Convert taumax-style sigmoid parameters to David's.}
\usage{
convert_taumax_to_David(taumax, benefitCurvatureAtTaumax, mmf)
}
\arguments{
\item{taumax}{maximizing total good}

\item{benefitCurvatureAtTaumax}{the curvature of the benefit
function at \code{taumax}}

\item{mmf}{maximal marginal fitness}
}
\description{
Converts the maximizing total good \code{taumax}, the curvature of the
benefit function at \code{taumax}, \code{benefitCurvatureAtTaumax},
and the maximal marginal fitness \code{mmf} to David's sigmoid
parametrization.
}
\note{
See easierMS appendix A.3 for derivation.
}
\examples{
pDavidV1 <- convert_Cornforth_to_David(a=100, b=0.2, beta=4.76, k=10, d=1)
pTaumax <- convert_David_to_taumax(A=pDavidV1[["A"]], K=pDavidV1[["K"]], L = pDavidV1[["L"]])
pDavidV2 <- convert_taumax_to_David(taumax=pTaumax[["taumax"]], benefitCurvatureAtTaumax=pTaumax[["benefitCurvatureAtTaumax"]], mmf=pTaumax[["mmf"]])
}
\references{
\insertRef{Corn+12}{evolvr}
}
