% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{cost}
\alias{cost}
\title{Cost function for \code{snowdrift.game}}
\usage{
cost(model, x)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{x}{focal individual's contribution}
}
\description{
Cost function for \code{snowdrift.game}
}
