% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{create_snowdrift_game}
\alias{create_snowdrift_game}
\alias{snowdrift.game}
\title{Create a \code{snowdrift.game} model}
\usage{
create_snowdrift_game(
  type = "tanh",
  nPlayers = 2,
  nGroups = 5,
  nRepetitions = 1,
  ...
)
}
\arguments{
\item{type}{character string indicating type of fitness function
(e.g., \code{"tanh"}, \code{"quadratic"})}

\item{nPlayers}{number of players in a single group}

\item{nGroups}{number of groups into which the population can be divided}

\item{nRepetitions}{number of repetitions of the game to be played between reproductive events (default=1)}

\item{...}{further arguments passed to
\code{\link{setup_gerf_snowdrift_game}},
\code{\link{setup_tanh_snowdrift_game}} or
\code{\link{setup_quadratic_snowdrift_game}}}

\item{simParms}{named list of simulation parameters}
}
\value{
a \code{snowdrift.game} object, which is a list containing
    many parameters, function definitions, and properties of the
    game (e.g., singular strategies).
}
\description{
A \strong{\emph{snowdrift game}} is a public goods game in which
the fitness of a focal individual can be written \deqn{W(x,\tau) =
B(\tau) - C(x) ,} where \eqn{x} is the focal individual's
contribution and \eqn{\tau} is the total group contribution.
}
\note{
The total population size is \code{Npop = nPlayers * nGroups}.
}
\examples{
sgtanh <- create_snowdrift_game()
summary(sgtanh)
plot(sgtanh)
sg <- create_snowdrift_game(type="gerf", nPlayers=4, nGroups=2)
summary(sg)
plot(sg)

}
\seealso{
\code{\link{setup_tanh_snowdrift_game}},
    \code{\link{setup_gerf_snowdrift_game}},
    \code{\link{setup_quadratic_snowdrift_game}},
    \code{\link{set_simParms}}
}
