% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{critical_nGroups}
\alias{critical_nGroups}
\title{Critical number of groups}
\usage{
critical_nGroups(model, Npop = model$parms$Npop)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{Npop}{population size}
}
\description{
The minimum number of groups for which a cooperative ESSN exists
for the given population size.
}
