% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{critical_nPlayers}
\alias{critical_nPlayers}
\title{Critical number of players}
\usage{
critical_nPlayers(model, Npop = model$parms$Npop)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{Npop}{population size}
}
\description{
The maximum number of players for which a cooperative ESSN exists
for the given population size.
}
