% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{critical_Npop}
\alias{critical_Npop}
\title{Critical population size}
\usage{
critical_Npop(model, nPlayers = model$parms$nPlayers)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{nPlayers}{number of players}
}
\description{
The minimum population size for which a cooperative ESSN exists for
the given number of players.
}
