% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{critical_Npop_from_nGroups}
\alias{critical_Npop_from_nGroups}
\title{Critical population size}
\usage{
critical_Npop_from_nGroups(model, nGroups = model$parms$nGroups)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{nGroups}{number of groups}
}
\description{
The maximum population size for which a cooperative ESSN exists for
the given number of groups.
}
