% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{dbenefit}
\alias{dbenefit}
\title{\emph{n}th derivative of benefit function for \code{snowdrift.game}}
\usage{
dbenefit(model, tau, nderiv = 1)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{tau}{total group contribution}
}
\description{
\emph{n}th derivative of benefit function for \code{snowdrift.game}
}
\details{
\code{nderiv=0} yields the benefit function itself
}
\seealso{
\code{\link{benefit}}
}
