% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{dcost}
\alias{dcost}
\title{\emph{n}th derivative of cost function for \code{snowdrift.game}}
\usage{
dcost(model, x, nderiv = 1)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{x}{focal individual's contribution}
}
\description{
\emph{n}th derivative of cost function for \code{snowdrift.game}
}
