% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{draw_tangent}
\alias{draw_tangent}
\title{Draw the tangent to a curve at a given point}
\usage{
draw_tangent(x, y, slope, width = 1, ...)
}
\arguments{
\item{x, y}{coordinates}

\item{slope}{slope of tangent line}

\item{width}{length of segment centered at the point \eqn{(x,y)}
projected onto \eqn{x} axis}

\item{...}{further \link{graphical parameters}}
}
\description{
Draw the tangent to a curve at a given point
}
