% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{elegend}
\alias{elegend}
\title{Default legend method for \pkg{evolvr} objects}
\usage{
elegend(object, x, y = NULL, ...)
}
\arguments{
\item{x, y}{the x and y co-ordinates to be used to position the legend.
    They can be specified by keyword or in any way which is accepted by
    \code{\link{xy.coords}}: See \sQuote{Details}.}
}
\description{
Default legend method for \pkg{evolvr} objects
}
\seealso{
\code{\link[graphics]{legend}}
}
