% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{elegend.simParms}
\alias{elegend.simParms}
\title{Legend for \code{simParms} objects}
\usage{
\method{elegend}{simParms}(sp, x = "left", y = NULL, iRes = 1, ...)
}
\arguments{
\item{sp}{a \code{\link[=set_simParms]{simParms}} object}

\item{iRes}{index of desired resident strategy}

\item{...}{further arguments passed to \code{\link{legend.default}}}
}
\description{
Legend for \code{simParms} objects
}
