% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{elegend.snowdrift.game}
\alias{elegend.snowdrift.game}
\title{\code{elegend} S3 method for \code{snowdrift.game} objects}
\usage{
\method{elegend}{snowdrift.game}(model, x = "topleft", y = NULL, digits = 4, ...)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{...}{further arguments passed to \code{\link{legend}}}
}
\description{
\code{elegend} S3 method for \code{snowdrift.game} objects
}
