% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{evolve_to_fixation}
\alias{evolve_to_fixation}
\title{Evolve a population until fixation.}
\usage{
evolve_to_fixation(model, pop, maxGen, selectionProcess)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{pop}{population matrix (columns: \code{strategy} and
\code{fitness})}

\item{maxGen}{maximum number of generations to evolve}

\item{selectionProcess}{selection process (\code{"WF"} = Wright-
Fisher, \code{"M"} = Moran, \code{"BD"} = birth-death)}
}
\value{
a list containing a logical flag (\code{mutantFixed}) and
    the number of generations that were computed (\code{fixTime})
}
\description{
Evolve a population until fixation.
}
