% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{evolve_with_mutation}
\alias{evolve_with_mutation}
\title{Evolve a population under selection and mutation}
\usage{
evolve_with_mutation(
  model,
  mutProb = 0.01,
  mutSD = 0.005,
  mutRange = c(-Inf, Inf),
  xMin = 0,
  xMax = 6,
  xInitMean = 5,
  xInitSD = mutSD,
  nGen = 1000,
  nSaveTimes = nGen,
  writeInterval = 10,
  saveTimes = NULL
)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{mutProb}{probability that a mutation occurs}

\item{mutSD}{standard deviation for normal distribution with zero
mean, used to shift offspring strategy if a mutation occurs}

\item{mutRange}{the minimal and maximal magnitudes of mutations
(=c(-Inf,Inf) by default)}

\item{xMin, xMax}{extreme values of playable strategies (boundaries
of strategy space)}

\item{xInitMean}{mean of a normally distributed set of initial
strategies for the population}

\item{xInitSD}{standard deviation of normally distributed initial
strategies with mean \code{xInitMean}}

\item{nGen}{number of generations to simulate}

\item{nSaveTimes}{number of times in \code{saveTimes}}

\item{writeInterval}{number of \code{saveTimes} between writing to disk}

\item{saveTimes}{generations at which to save population for later
analysis}
}
\value{
A \code{\link{matrix}}, with one column for each individual
    in the population, and one row for each saved generation.  The
    matrix entries are the strategies of the individuals.
}
\description{
Evolve a population under selection and mutation
}
\details{
By default, \code{nSaveTimes} generations are saved,
    equally spaced between \code{0} and \code{nGen}, unless
    \code{nGen < nSaveTimes}, in which case every generation is
    saved.  If a vector of \code{saveTimes} is given then
    \code{nSaveTimes} is ignored.
}
\seealso{
\code{\link{selection_step_WF}}, \code{\link{selection_step_M}},
    \code{\link{selection_step_BD}}
}
