% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{evolvr-package}
\alias{evolvr}
\alias{evolvr-package}
\title{The \pkg{evolvr} package}
\description{
\pkg{evolvr} is a package for studying evolutionary games in finite
populations.
}
\details{
This is the version of the `evolvr` package used to make the
figures in Molina and Earn (2024, Proc. R. Soc. A, 20230305,
\url{https://doi.org/10.1098/rspa.2023.0305}).  The
latest version of the package is available on github at
\url{https://github.com/davidearn/evolvr} and might also be
available on CRAN.

This version does not have an associated vignette, but we plan to
include a vignette in future versions, which should be possible to
find via \code{browseVignettes("evolvr")}.
}
\author{
\strong{Maintainer}: David Earn \email{earn@math.mcmaster.ca}

Authors:
\itemize{
  \item Chai Molina \email{chai.molina@gmail.com}
  \item Alexei Kuzmin \email{kuzmia1@mcmaster.ca}
}

}
