% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{fitness}
\alias{fitness}
\title{Fitness function for \code{snowdrift.game}}
\usage{
fitness(model, x, tau)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{x}{focal individual's contribution}

\item{tau}{total group contribution}
}
\description{
Fitness function for \code{snowdrift.game}
}
