% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{gerf}
\alias{gerf}
\title{Generalized error function via incomplete gamma function}
\usage{
gerf(x, n)
}
\arguments{
\item{x}{argument}

\item{n}{order of generalized erf; \eqn{n=2} yields the normal erf}
}
\description{
\deqn{E_n(x) = (1/\sqrt\pi) * \Gamma(n) * ( \Gamma(1/n) - \Gamma(1/n, x^n) ) }
See \url{https://en.wikipedia.org/wiki/Error_function#Generalized_error_functions}.
Unlike \eqn{E_n}, \code{gerf} is normalized so that its maximum
is \code{1} regardless of the value of \code{n}.
}
\examples{
x <- seq(-3,3,length=1000)
g2 <- gerf(x,2)
g4 <- gerf(x,4)
g8 <- gerf(x,8)
g16 <- gerf(x,16)
g32 <- gerf(x,32)
y <- cbind(g2,g4,g8,g16,g32)
plot(NA, NA, xlim=range(x), ylim=c(-1,1), type="n", bty="L", las=1,
     xlab=expression(x), ylab="gerf(x,n)",
     main="Generalized Error Function")
grid()
matlines(x,y, lty=1, lwd=2)
plot(NA, NA, xlim=c(0.75,1.25), ylim=c(0.8,1), type="n", bty="L", las=1,
     xlab=expression(x), ylab="gerf(x,n)",
     main="Generalized Error Function (zoomed in)")
grid()
matlines(x,y, lty=1, lwd=2)

}
\seealso{
\code{\link[gsl]{gamma_inc}}, \code{\link{gerf_norm}}
}
