% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{get_deriv_strings}
\alias{get_deriv_strings}
\title{Get symbolic derivatives as strings}
\usage{
get_deriv_strings(string, var = "x", nderiv = 0:2)
}
\arguments{
\item{string}{function expression as character string (\code{f}
argument of \code{\link[Deriv]{Deriv}})}

\item{var}{variable name (\code{x} argument of
\code{\link[Deriv]{Deriv}})}

\item{nderiv}{vector of desired derivative orders (\code{nderiv}
argument of \code{\link[Deriv]{Deriv}})}
}
\value{
named list of character strings
}
\description{
Get symbolic derivatives as strings
}
\examples{
get_deriv_strings("a*x^4")

}
