% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{get_singular_strategy}
\alias{get_singular_strategy}
\title{Get singular strategy}
\usage{
get_singular_strategy(model, defaultStrategy = 1)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}
}
\description{
Either the infinite population singular strategy \code{XstarInf} if
it exists or the \code{defaultStrategy} otherwise.
}
