% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{initialize_population}
\alias{initialize_population}
\title{Initialize population with residents and mutants}
\usage{
initialize_population(model, resident, mutant, nMut)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{resident}{resident strategy}

\item{mutant}{mutant strategy}

\item{nMut}{number of mutants}
}
\value{
a \code{\link{population_matrix}}
}
\description{
Population is represented by an array of numbers, with rows
representing individual agents.  Columns are \code{"strategy"} and
\code{"fitness"}.  Mutants are randomly distributed among groups.
}
\note{
Agent in row \code{i} is considered to be in group
    \code{ceiling(i/nPlayers)}, where groups are numbered
    \code{1:nGroups} and rows are numbered \code{1:Npop}.
}
\seealso{
\code{\link{initialize_population_randomly}}
}
