% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{initialize_population_randomly}
\alias{initialize_population_randomly}
\title{Initialize population via a Normal distribution of strategies}
\usage{
initialize_population_randomly(
  model,
  xMean,
  xSD,
  mutRange = c(-Inf, Inf),
  xMin,
  xMax
)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{xMean}{mean strategy}

\item{xSD}{standard deviation of strategy distribution}

\item{xMin, xMax}{extreme values of playable strategies (boundaries
of strategy space)}
}
\value{
a \code{\link{population_matrix}}
}
\description{
Initialize population via a Normal distribution of strategies
}
\seealso{
\code{\link{initialize_population}}, \code{\link{rnorm}}
}
