% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{is_nsg}
\alias{is_nsg}
\title{Is this a Natural Snowdrift Game?}
\usage{
is_nsg(model)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}
}
\description{
Is this a Natural Snowdrift Game?
}
