% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{legend_location}
\alias{legend_location}
\title{Legend location}
\usage{
legend_location(x, y = NULL, xjust = 0, yjust = 1)
}
\arguments{
\item{x, y}{the x and y co-ordinates to be used to position the legend.
    They can be specified by keyword or in any way which is accepted by
    \code{\link{xy.coords}}: See \sQuote{Details}.}

\item{xjust}{how the legend is to be justified relative to the legend
    x location.  A value of 0 means left justified, 0.5 means centered
    and 1 means right justified.}

\item{yjust}{the same as \code{xjust} for the legend y location.}
}
\value{
list of revised versions of arguments suitable for passing
    to \code{\link{legend}}
}
\description{
This function adds additional named locations to those allowed by
    \code{\link{legend}}.  In addition to what \code{\link{legend}}
    allows, you can set \code{x} to any of \code{"middleleft"},
    \code{"middleright"}, \code{"topmiddle"},
    \code{"bottommiddle"}.
}
\note{
It turns out \code{\link{legend}} already has these options
    built in, but called \code{"top"}, \code{"bottom"} \emph{etc.}
    See \code{example(legend) (one of the last examples)}.
}
\examples{
x <- 1:100
y <- rnorm(100)
plot(x,y)
legloc <- legend_location("middleleft")
legend(legloc$x, legloc$y, xjust=legloc$xjust, yjust=legloc$yjust, legend="middle left legend")
legloc <- legend_location("middleright")
legend(legloc$x, legloc$y, xjust=legloc$xjust, yjust=legloc$yjust, legend="middle right legend")
legloc <- legend_location("topmiddle")
legend(legloc$x, legloc$y, xjust=legloc$xjust, yjust=legloc$yjust, legend="top middle legend")
legloc <- legend_location("bottommiddle")
legend(legloc$x, legloc$y, xjust=legloc$xjust, yjust=legloc$yjust, legend="bottom middle legend")
}
\seealso{
\code{\link{legend_simParms}}, \code{\link{legend_snowdrift_game}}
}
