% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{legend_run_info}
\alias{legend_run_info}
\title{Run info legend}
\usage{
legend_run_info(x, xwhere = "topright", ywhere = NULL, digits = 4, ...)
}
\arguments{
\item{...}{further arguments passed to \code{\link{legend}}}
}
\description{
Run info legend
}
