% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{make_fitness_positive}
\alias{make_fitness_positive}
\title{Make fitness positive}
\usage{
make_fitness_positive(pop, fitnessMin)
}
\arguments{
\item{pop}{population matrix: named columns (\code{fitness} and
\code{strategy}) and one row for each agent}
}
\description{
Probabilities of generating an offspring are supposed to be
proportional to fitness, but some fitnesses are negative. This
makes them positive.
}
\details{
It is important the \code{fitnessMin} be positive.  If it
    is zero, then mutants will never fix if they have unadjusted
    fitness less than residents.
}
\seealso{
\code{\link{selection_step_WF}}
}
