% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sigmoid.R
\name{mean_fitness_slope_erf}
\alias{mean_fitness_slope_erf}
\title{Mean fitness slope (between extrema) as a function of \code{mmf}, for erf-style sigmoid}
\usage{
mean_fitness_slope_erf(mmf)
}
\arguments{
\item{mmf}{maximal marginal fitness}
}
\description{
\eqn{\alpha} is the ratio between the difference between the
heights of the two fitness extrema and the difference between the
maximizing and minimizing total goods, calculated for erf-style sigmoid
}
