% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{mean_fitness_slope_gerf}
\alias{mean_fitness_slope_gerf}
\title{Mean fitness slope with sigmoid via gerf}
\usage{
mean_fitness_slope_gerf(mmf, nben)
}
\arguments{
\item{mmf}{maximal marginal fitness}

\item{nben}{the index of the generalized error function is \code{2*nben}
(FIX: think of a better name than \code{nben} for this...)}
}
\description{
Mean fitness slope with sigmoid via generalized error functions
}
\examples{
nvec <- 1:50
mfsg <- mean_fitness_slope_gerf(mmf=1,nben=nvec)
plot(NA, NA, xlim=range(nvec), ylim=c(0,1), bty="L", las=1,
     xlab="nben", ylab="Mean fitness slope")
abline(h=1, col="grey")
lines(nvec, mfsg, type="o", pch=21, bg="darkred", cex=0.5)
nn <- length(nvec)
message("For n=", nvec[nn], ", mfsg=", mfsg[nn])
}
\seealso{
\code{\link{gerf}}, \code{\link{benefit}},
    \code{\link{fitness}},
    \code{\link{setup_gerf_snowdrift_game}},
    \code{\link{mean_fitness_slope_erf}},
    \code{\link{mean_fitness_slope_tanh}}
}
