% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{merge_fixationData}
\alias{merge_fixationData}
\title{Merge several \code{fixationData} objects}
\usage{
merge_fixationData(x)
}
\arguments{
\item{x}{a list of \code{fixationData} objects}
}
\description{
Merge several \code{fixationData} objects
}
