% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{mmf_fun}
\alias{mmf_fun}
\title{maximum marginal fitness for NSG}
\usage{
mmf_fun(model)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}
}
\description{
maximum marginal fitness for NSG
}
