% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{multiple_evolve_to_fixation}
\alias{multiple_evolve_to_fixation}
\title{Evolve each of a sequence of models to fixation}
\usage{
multiple_evolve_to_fixation(model, simParms, ...)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{simParms}{a \code{\link[=set_simParms]{simParms}} object}
}
\description{
Evolve each of a sequence of models to fixation
}
