% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{mutate_agent_strategies}
\alias{mutate_agent_strategies}
\title{Mutates strategies of a set of agents}
\usage{
mutate_agent_strategies(
  model,
  pop,
  whichAgentsMutate,
  mutProb,
  mutSD,
  mutRange,
  xMin,
  xMax
)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{whichAgentsMutate}{vector of agents whose strategies mutate}

\item{mutProb}{probability that a mutation occurs}

\item{mutSD}{standard deviation for normal distribution with zero
mean, used to shift offspring strategy if a mutation occurs}

\item{mutRange}{the minimal and maximal magnitudes of mutations
(=c(-Inf,Inf) by default)}

\item{xMin, xMax}{extreme values of playable strategies (boundaries
of strategy space)}
}
\description{
Mutates strategies of a set of agents
}
