% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{mutation_step_nonoverlapping}
\alias{mutation_step_nonoverlapping}
\title{Mutation step for simulations with non-overlapping generations
(e.g., Wright-Fisher).}
\usage{
mutation_step_nonoverlapping(model, pop, mutProb, mutSD, mutRange, xMin, xMax)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{mutProb}{probability that a mutation occurs}

\item{mutSD}{standard deviation for normal distribution with zero
mean, used to shift offspring strategy if a mutation occurs}

\item{mutRange}{the minimal and maximal magnitudes of mutations
(=c(-Inf,Inf) by default)}

\item{xMin, xMax}{extreme values of playable strategies (boundaries
of strategy space)}
}
\description{
Mutation step for simulations with non-overlapping generations
(e.g., Wright-Fisher).
}
