% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{mutation_step_WF}
\alias{mutation_step_WF}
\title{Mutation step for Wright-Fisher simulations}
\usage{
mutation_step_WF(model, pop, mutProb, mutSD, mutType = "norm", xMin, xMax)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{mutProb}{probability that a mutation occurs}

\item{mutSD}{standard deviation for normal distribution with zero
mean, used to shift offspring strategy if a mutation occurs}

\item{xMin, xMax}{extreme values of playable strategies (boundaries
of strategy space)}
}
\description{
Mutation step for Wright-Fisher simulations
}
