% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{my_grid}
\alias{my_grid}
\title{More flexible grid function}
\usage{
my_grid(
  nx = NULL,
  ny = nx,
  col = "lightgray",
  lty = "dotted",
  lwd = par("lwd"),
  eps.correct = 1,
  avoid.axes = TRUE,
  ...
)
}
\arguments{
\item{nx, ny}{number of cells of the grid in x and y direction.  When
    \code{NULL}, as per default, the grid aligns with the tick marks on
    the corresponding \emph{default} axis (i.e., tickmarks as computed by
    \code{\link[graphics]{axTicks}}).  When \code{\link{NA}}, no grid lines are
    drawn in the corresponding direction.}

\item{col}{character or (integer) numeric; color of the grid lines.}

\item{lty}{character or (integer) numeric; line type of the grid lines.}

\item{lwd}{non-negative numeric giving line width of the grid lines.}

\item{eps.correct}{integer code, one of \{0,1,2\}. If non-0, an
    \emph{epsilon correction} is made at the boundaries such that
    the result boundaries will be outside \code{range(x)}; in the
    \emph{small} case, the correction is only done if \code{eps.correct >= 2}.}

\item{...}{additional parameters passed to \code{\link{abline}}}
}
\description{
More flexible grid function
}
\note{
See the examples in the help for \code{\link{plot.window}}
}
\seealso{
\code{\link{grid}}
}
