% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{plot.evolutionData}
\alias{plot.evolutionData}
\title{Plot method for \code{evolutionData} objects}
\usage{
\method{plot}{evolutionData}(x, pch = ".", las = 1, col = "grey", ...)
}
\arguments{
\item{x}{an \code{evolutionData} object}
}
\description{
Plot method for \code{evolutionData} objects
}
