% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{plot.snowdrift.game}
\alias{plot.snowdrift.game}
\title{Plot method for \code{snowdrift.game} objects}
\usage{
\method{plot}{snowdrift.game}(
  model,
  add = FALSE,
  residentStrategy = get_singular_strategy(model),
  what = "all",
  xmin = 0,
  xmax = reasonable_xmax(model),
  xmax.group = NULL,
  ylim.fitness = NULL,
  show.main = TRUE,
  show.leftlab = FALSE,
  cex.leftlab = 2,
  leftlab.fitness = "Fitness",
  show.annotation = TRUE,
  draw.special.strategies = TRUE,
  draw.XstarN = TRUE,
  draw.XstarInf = TRUE,
  col.resident = "yellow",
  lwd.resident = 5,
  lty.resident = "dotted",
  label.resident = "resident strategy",
  cex.label.resident = 0.8,
  lty.XstarInf = "solid",
  lty.XstarN = "solid",
  col.XstarInf = "grey",
  col.XstarN = "black",
  draw.grid = TRUE,
  draw.benefit.tangents = FALSE,
  xlab.focal = "focal individual's contribution",
  xlab.total = "total group contribution",
  ylab.cost = "Fitness cost",
  ylab.benefit = "Fitness benefit",
  ylab.fitness = "Fitness",
  col.cost = "darkred",
  col.benefit = "darkblue",
  col.fitness = "darkgreen",
  xpd.cost = FALSE,
  xpd.benefit = NA,
  xpd.fitness = NA,
  lwd = 5,
  las = 1,
  xaxs = "i",
  bty = "L",
  ...
)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{residentStrategy}{contribution level of residents}

\item{what}{character vector: what to plot (\code{"cost"},
\code{"benefit"}, \code{"fitness"}, \code{"dcost"},
\code{"dbenefit"}, \code{"d2benefit"}, \code{"all"}, or any
combination)}

\item{xmin, xmax}{domain limits for focal individual's strategy}

\item{xmax.group}{domain limit for total group contribution}

\item{ylim.fitness}{upper limit for fitness plot}

\item{xlab.focal, xlab.total}{\eqn{x} axis labels for individual
vs. total group contributions}

\item{...}{other \code{\link{graphical parameters}}}
}
\description{
Plot method for \code{snowdrift.game} objects
}
