% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{plot_fixation_time}
\alias{plot_fixation_time}
\title{Plot fixation time as a function of mutant strategy}
\usage{
plot_fixation_time(
  x,
  xlim = NULL,
  ylim = NULL,
  bg = c(mutant = "darkred", resident = "darkblue"),
  cex = 1,
  las = 1,
  xlab = "Mutant strategy",
  ylab = " fixation time",
  bty = "L",
  ...
)
}
\arguments{
\item{x}{a \code{fixationData} object}

\item{...}{\code{\link{graphical parameters}}}
}
\description{
Plot fixation time as a function of mutant strategy
}
