% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{population_matrix}
\alias{population_matrix}
\title{Create a population matrix object}
\usage{
population_matrix(Npop)
}
\value{
\code{\link{matrix}} with \code{Npop} rows and three
    columns: \code{"ID"}, \code{"strategy"}, and \code{"fitness"}
}
\description{
Create a population matrix object
}
\examples{
population_matrix( 10 )
}
\seealso{
\code{\link{initialize_population}},
    \code{\link{initialize_population_randomly}}
}
