% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{reasonable_xmax}
\alias{reasonable_xmax}
\title{Compute reasonable upper limit for fitness function plots}
\usage{
reasonable_xmax(model, xmax.default = 2)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}
}
\description{
What is reasonable depends on whether an ESS exists.
If not, we used \code{xmax.default}.
If so, we use the ESS multiplied by \code{xmax.default}.
}
