% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_interface_funs.R
\name{restore_rng}
\alias{restore_rng}
\title{Restore random number generator seed from file}
\usage{
restore_rng(savefile)
}
\description{
Restore random number generator seed from file
}
\note{
Taken from \href{http://stackoverflow.com/questions/13997444/print-the-current-random-seed-so-that-i-can-enter-it-with-set-seed-later}{Ben Bolker's answer on stackoverflow}.
}
