% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_interface_funs.R
\name{save_rng}
\alias{save_rng}
\title{Save random number generator seed to file}
\usage{
save_rng(savefile = tempfile())
}
\description{
Save random number generator seed to file
}
\note{
Taken from \href{http://stackoverflow.com/questions/13997444/print-the-current-random-seed-so-that-i-can-enter-it-with-set-seed-later}{Ben Bolker's answer on stackoverflow}.
}
