% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{save_simulation_progress}
\alias{save_simulation_progress}
\title{Save simulation progress}
\usage{
save_simulation_progress(x, iRes = NA, iMut = NA)
}
\arguments{
\item{x}{a \code{fixationData} object}

\item{iRes}{index of resident strategy}

\item{iMut}{index of mutant strategy}
}
\description{
This function is called by
\code{\link{multiple_evolve_to_fixation}} to save data periodically
during a sequence of simulations, and to write a plain text
progress file to help monitor a job's status.
}
\note{
FIX: this function should probably return a status flag,
    indicating whether saving was successful.
}
