% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{selection_step_BD}
\alias{selection_step_BD}
\title{Advance by one generation of Birth-Death process}
\usage{
selection_step_BD(model, pop, fitnessMin = 1)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{pop}{population matrix}
}
\description{
Constructs new generation based on the transition probabilities of
a birth-death process, which are in turn calculated from the mean
fitnesses of mutants and residents in the current generation.  From
state i, the state can either increase to i+1, decrease to i-1, or
remain i.  This selection process is not defined for more than two
traits in the population.  Note that the fitnesses of the current
generation are not calculated here and are assumed to be input in
\code{pop[,"fitness"]}, but the fitnesses of the next generation
are output in \code{pop[,"fitness"]}.
}
