% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{selection_step_M}
\alias{selection_step_M}
\title{Advance population by one generation using Moran selection process}
\usage{
selection_step_M(model, pop, fitnessMin = 1)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{pop}{population matrix: named columns (\code{fitness} and
\code{strategy}) and one row for each agent}
}
\value{
population matrix with \code{hasFixed} attribute to
    indicate whether or not the population has reached fixation.
}
\description{
Constructs new generation based on frequency-dependent Moran
process
}
\details{
Assumes fitnesses are calculated. Agent is chosen to
    reproduce (with a probability proportional to its fitness) and
    another agent is chosen from the population (with uniform
    probability) to die. The strategy of the dying agent is
    replaced by the strategy of the reproducing one. Randomizes
    group allocation in the population, and calculates individual
    fitnessess based on payoffs from games played in these groups.
}
