% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{selection_step_WF}
\alias{selection_step_WF}
\title{Advance population by one generation using Wright-Fisher selection process}
\usage{
selection_step_WF(model, pop, fitnessMin = 1)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{pop}{population matrix: named columns (\code{fitness} and
\code{strategy}) and one row for each agent}
}
\value{
population matrix with \code{hasFixed} attribute to
    indicate whether or not the population has reached fixation.
}
\description{
Advance population by one generation using Wright-Fisher selection process
}
\details{
FIX: this is mostly old comments.
Assumes fitnesses have been calculated and are available
    as the second column of the \code{pop} matrix. Next generation
    is constructed using N independent draws from the pool of
    strategies currently in the population. The probability of
    drawing strategy x in the pool for any one of the next
    generation's agents is proportional to the total fitness from
    all the agents playing x in the current population (that is, to
    the mean fitness of x and to the number of agents playing
    x). Randomizes group allocation in population, and calculates
    individual fitnessess based on payoffs from games played in
    these groups.
}
