% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{set_simParms}
\alias{set_simParms}
\title{Set simulation parameters}
\usage{
set_simParms(
  model,
  maxGen = 100,
  nReplicates = 100,
  mutantMin = 1e-06,
  mutantMax = 5,
  nMutantStrategies = 100,
  residentMin = 0,
  residentMax = model$parms$XstarInf,
  nResidentStrategies = 2,
  selectionProcess = "WF"
)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{maxGen}{maximum allowed generations}

\item{nReplicates}{number of replicate simulations}

\item{nMutantStrategies}{number of mutant strategies}

\item{nResidentStrategies}{number of resident strategies; enables
multiple simultaneous jobs (should be \code{<=} number of
processors on machine).  FIX: not currently set to run in
parallel}

\item{selectionProcess}{character string indicating which selection
process to use: Wright-Fisher (\code{"WF"}), Moran
(\code{"M"}), or birth-death (\code{"BD"})}
}
\value{
a \code{simParms} object, which is just a list of the
    parameters and their values
}
\description{
Set simulation parameters
}
