% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{setup_gerf_snowdrift_game}
\alias{setup_gerf_snowdrift_game}
\title{Set up snowdrift game with sigmoidal benefit function}
\usage{
setup_gerf_snowdrift_game(
  mmf = 1,
  tauTurn = 1,
  L = 1,
  nben = 1,
  c1 = 1,
  c2 = 0
)
}
\arguments{
\item{mmf}{maximal marginal fitness}

\item{tauTurn}{inflection point}

\item{L}{benefit limit: the limit of \eqn{B(\tau)} as \eqn{\tau}
approaches \eqn{\infty}}

\item{nben}{the index of the generalized error function is \code{2*nben}
(FIX: think of a better name than \code{nben} for this...)}

\item{c1}{coefficient of linear term of cost function}

\item{c2}{coefficient of quadratic term of cost function}
}
\value{
list containing benefit and cost functions (as strings) and
    their parameters as lists
}
\description{
This function is called by \code{\link{create_snowdrift_game}} as
one of the basic setup options.  The benefit function is designed
using \code{\link[=gerf]{generalized error functions}}.
The cost function is quadratic in principle, \deqn{C(x) = c_1 x +
c_2 x^2 ,} but the default parameter values yield \eqn{C(x) = x}
and hence a \strong{\emph{natural snowdrift game}} (NSG) in the
sense of Molina and Earn (2018).
}
\details{
The returned \code{benefitString} is a character string
    giving the full expression for the fitness \emph{benefit} as a
    function of the total group contribution \code{tau}.
    Similarly, the returned \code{costString} is a character string
    giving the full expression for the fitness \emph{cost} as a
    function of the focal individual's contribution \code{x}.
    These strings are used in \code{\link{create_snowdrift_game}}
    to construct functions and their derivatives.
}
\seealso{
\code{\link{gerf}}
}
