% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{setup_quadratic_snowdrift_game}
\alias{setup_quadratic_snowdrift_game}
\title{Set up \code{snowdrift.game} fitness function based on quadratic
cost and benefit functions}
\usage{
setup_quadratic_snowdrift_game(
  Npop = 10,
  nPlayers = 2,
  nu = 3/2,
  xi = Npop,
  b1 = 0,
  b2 = 1,
  c1 = 2,
  c2 = c2_fun(Npop, nPlayers, nu, xi)
)
}
\arguments{
\item{Npop}{population size}

\item{nPlayers}{number of players}

\item{nu}{non-negative real number that indexes the examples}

\item{b1, b2, c1, c2}{coefficients of linear and quadratic terms in
benefit and cost functions (defaults are chosen to yield the
class of examples used in the Invasion-Fixation paper)}
}
\value{
list containing benefit and cost functions (as strings) and
    their parameters as lists
}
\description{
This is the class of snowdrift games used in the Invasion-Fixation
paper to illustrate the results.
}
\note{
The values of the arguments that are used to define `c2` by
    default (i.e., `N`, `nGroups`, `nu` and `xi`) are saved as attributes of
    the `costParms` list.
}
\seealso{
\code{\link{c2_fun}}
}
