% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{string_to_expr}
\alias{string_to_expr}
\title{Convert strings to expressions}
\usage{
string_to_expr(stringList)
}
\arguments{
\item{stringList}{a list of character strings}
}
\description{
Convert strings to expressions
}
