% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{summary.simParms}
\alias{summary.simParms}
\title{Summary method for \code{simParms} objects}
\usage{
\method{summary}{simParms}(x)
}
\arguments{
\item{x}{a \code{\link[=set_simParms]{simParms}} object}
}
\description{
Summary method for \code{simParms} objects
}
