% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{summary.snowdrift.game}
\alias{summary.snowdrift.game}
\title{Summary method for \code{snowdrift.game} objects}
\usage{
\method{summary}{snowdrift.game}(model)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}
}
\description{
Summary method for \code{snowdrift.game} objects
}
