% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{time_string}
\alias{time_string}
\title{Human-friendly time string}
\usage{
time_string(seconds, sigdig = 4)
}
\arguments{
\item{seconds}{a real number}

\item{sigdig}{number of significant digits to retain}
}
\description{
Human-friendly time string
}
