% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{total_group_contribution}
\alias{total_group_contribution}
\title{Total group contribution in \code{snowdrift.game}}
\usage{
total_group_contribution(model, resident, focal)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}
}
\description{
Total group contribution in \code{snowdrift.game}
}
