% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{XstarInf_fun}
\alias{XstarInf_fun}
\title{Infinite population singular strategy}
\usage{
XstarInf_fun(model, nPlayers = model$parms$nPlayers)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}
}
\description{
Infinite population singular strategy
}
