% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{XstarN_fun}
\alias{XstarN_fun}
\title{Finite population singular strategy}
\usage{
XstarN_fun(
  model,
  pm = +1,
  nPlayers = model$parms$nPlayers,
  Npop = model$parms$Npop
)
}
\arguments{
\item{model}{a \code{\link{snowdrift.game}} object}

\item{pm}{plus or minus: \code{+1/-1} for stable/unstable singular strategy}

\item{nPlayers}{number of players}

\item{Npop}{population size}
}
\description{
Finite population singular strategy
}
\examples{
sg <- create_snowdrift_game(type="gerf")
XstarN_fun(sg)
XstarN_fun(sg, -1)
}
