% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdrift.R
\name{XstarN_gerf}
\alias{XstarN_gerf}
\title{Finite population singular strategy for gerf sigmoid}
\usage{
XstarN_gerf(tauTurn, nben, L, mmf, pm, Npop, nPlayers)
}
\arguments{
\item{tauTurn}{inflection point}

\item{nben}{the index of the generalized error function is \code{2*nben}
(FIX: think of a better name than \code{nben} for this...)}

\item{L}{benefit limit: the limit of \eqn{B(\tau)} as \eqn{\tau}
approaches \eqn{\infty}}

\item{mmf}{maximal marginal fitness}

\item{pm}{plus or minus: \code{+1/-1} for stable/unstable singular strategy}

\item{Npop}{population size}

\item{nPlayers}{number of players}
}
\description{
Finite population singular strategy for gerf sigmoid
}
